ValKit/bs3kit: Warning fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@97985 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/ValidationKit/bootsectors/bs3kit/bs3-cmn-SelLnkPtrToFlat.c
@@ -41,7 +41,7 @@
 BS3_CMN_DEF(uint32_t, Bs3SelLnkPtrToFlat,(void BS3_FAR *pvLnkPtr))
 {
 #if ARCH_BITS == 16
-    return Bs3SelRealModeDataToFlat(pvLnkPtr);
+    return Bs3SelRealModeDataToFlat((uint32_t)pvLnkPtr);
 #else
     return (uint32_t)(uintptr_t)pvLnkPtr;
 #endif
